refactor(tasks): move Go2W shared base [skip ci]
