--- src/unilab/envs/locomotion/go2w/__init__.py
@@ -1,1 +1,0 @@
-"""Legacy Go2W shared base pending migration to :mod:`unilab.tasks`."""
